+++ src/Render/NeoInputRenderer.php
@@ -97,11 +97,13 @@ class NeoInputRenderer
      */
     public function selectBox($el, array $attrs, array $options): string
     {
+        $prompt = $el instanceof SelectBox ? $el->getPrompt() : null;
         return $this->block('selectBox', [
             'attrs' => $attrs,
             'options' => $el->getOptions() + $options,
             'items' => $el->getItems(),
             'value' => $el->getValue(),
+            'prompt' => $prompt,
         ]);
     }